Fully tested Visual Studio 2015
Multile files are removed and modified modify to fully support
Windows without any warning.

* Use CMake POLICY CMP0054 to suport `if ("${...}" STREQUAL MSVC)`
* Specify project language: C
* Make FindCheck.cmake to fully support Windows and Linux
* Remove iconv and write few code to transform utf16le to utf8
* Add instructions of how to build project in Windows
* Make OpenSSL and ZLIB required
* Add more GCC warning flags:
  `-Wextra -Wconversion -Wshadow -Wwrite-strings`
＊Add MSVC warning flags and ignore some of it:
  `/wd4204` # C99
  `/wd4710` # C89: inline function
  `/wd4820` # C89: struct padding
  `/wd4996` # C89: sprintf & fopen
  `/WX /Wall`
* Install shared library wz to bin
* Install header file.h to include/wz
* Add const keyword in `(const uint16_t *) "\x00\xff"`
* Use uint32_t instead of size_t in wz* structures and their functions
* Make length of string non fixed to allow compiler to add '\0'
* Remove VLA to support Visual Studio
* Cast non fixed size to fixed when doing bitwise operations
* Remove not used arguments of functions
* Make different variable different name in same function
* Disable MSVC struct padding warning in `file.h`
* Create `check_portable.h` to define `pid_t` as `int` in Windows
* Create `wrap.h` to allow wrap `malloc` and `free` in unit test

+++ tests/encode.c
@@ -18,7 +18,7 @@ main(void) {
   wzfile file = {.head = {.start = start}, .ver = ver};
   wzaddr addr = {.pos = position};
   for (uint64_t encoded = 0; encoded <= UINT32_MAX; encoded++) {
-    addr.val = encoded;
+    addr.val = (uint32_t) encoded;
     wz_decode_addr(&addr, &file);
     if (addr.val == address) {
       printf("encoded address: 0x%08"PRIx32"\n", (uint32_t) encoded);
